requirements.txt
aisuite==0.1.7
annotated-types==0.7.0
anthropic==0.30.1
anyio==4.8.0
attrs==24.3.0
beautifulsoup4==4.12.3
blinker==1.9.0
bs4==0.0.2
cachetools==5.5.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
cohere==5.13.6
decorator==4.4.2
distro==1.9.0
fastavro==1.10.0
filelock==3.16.1
Flask==3.1.0
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
fsspec==2024.12.0
google-ai-generativelanguage==0.6.10
google-api-core==2.24.0
google-api-python-client==2.156.0
google-auth==2.37.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
google-genai==1.0.0
google-generativeai==0.8.3
googleapis-common-protos==1.66.0
groq==0.9.0
grpcio==1.68.1
grpcio-status==1.68.1
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.27.2
httpx-sse==0.4.0
huggingface-hub==0.27.1
idna==3.10
imageio==2.36.1
imageio-ffmpeg==0.5.1
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.5
jiter==0.8.2
joblib==1.4.2
MarkupSafe==3.0.2
moviepy==1.0.3
numpy==2.2.1
oauthlib==3.2.2
openai==1.59.6
outcome==1.3.0.post0
packaging==24.2
pandas==2.2.3
parameterized==0.9.0
pillow==10.4.0
pluggy==1.5.0
proglog==0.1.10
proto-plus==1.25.0
protobuf==5.29.2
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.10.4
pydantic_core==2.27.2
pygame==2.6.1
pyparsing==3.2.0
PySocks==1.7.1
pytest==8.3.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
selenium==4.27.1
setuptools==75.6.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
SQLAlchemy==2.0.36
tokenizers==0.21.0
tqdm==4.67.1
trio==0.28.0
trio-websocket==0.11.1
types-requests==2.32.0.20241016
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.3.0
websocket-client==1.8.0
Werkzeug==3.1.3
wsproto==1.2.0
